projeto_extracao/requirements.txt
streamlit
pandas
plotly
sqlalchemy
pymysql
python-dotenv
matplotlib
seaborn
scikit-learn
lxml
openpyxl
